+++ src/Resources/contao/dca/tl_saison.php
@@ -14,10 +14,11 @@ declare(strict_types=1);
 
 
 use Contao\DataContainer;
+use Contao\DC_Table;
 
 $GLOBALS['TL_DCA']['tl_saison'] = [
     'config' => [
-        'dataContainer' => 'Table',
+        'dataContainer' => DC_Table::class,
         'enableVersioning' => true,
         'sql' => [
             'keys' => [